Fixed the CardItem on use method Added More Translation Support

+++ src/main/java/net/hanas_cards/component/CardComponent.java
@@ -5,17 +5,19 @@ import java.util.List;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import net.minecraft.network.PacketByteBuf;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.world.World;
 
 
 public class CardComponent {
-    private final int grading; // Field to store the grading of the card
-    private final String description; // Field to store the description of the card
-    private final String[] tags; // Field to store tags associated with the card
+    private final float grading;
+    private final String description;
+    private final String[] tags;
 
     public static final Codec<CardComponent> CODEC = RecordCodecBuilder.create(
             instance -> instance.group(
-                    Codec.INT.fieldOf("grading").forGetter(CardComponent::getGrading),
+                    Codec.FLOAT.fieldOf("grading").forGetter(CardComponent::getGrading),
                     Codec.STRING.fieldOf("description").forGetter(CardComponent::getDescription),
                     Codec.STRING.listOf().fieldOf("tags").forGetter(CardComponent::getTagList) // Modified getter here
             ).apply(instance, CardComponent::new)
@@ -24,14 +26,14 @@ public class CardComponent {
     public static final CardComponent DEFAULT = new CardComponent(0, "", List.of());
 
     // Constructor to initialize grading, description, and tags
-    public CardComponent(int grading, String description, List<String> tags) {
+    public CardComponent(float grading, String description, List<String> tags) {
         this.grading = grading; // Set the grading value
         this.description = description; // Set the description value
         this.tags = tags.toArray(new String[0]); // Convert List to array
     }
 
     // Getter for grading
-    public int getGrading() {
+    public float getGrading() {
         return grading;
     }
